ml/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# LLM - OpenAI-compatible (Moonshot / Kimi)
openai>=1.40.0

# Embeddings - Local sentence-transformers
sentence-transformers==3.3.1
torch==2.5.1

# Vector DB - ChromaDB
chromadb==0.5.23

# Graph DB - Neo4j
neo4j==5.27.0

# Parsing
tree-sitter==0.24.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-cpp==0.23.4
tree-sitter-java==0.23.5

# Utilities
python-dotenv==1.0.1
difflib2==0.0.1
